Make Wirecast conform to adopted VSA pattern
