+++ LivingMessiah.Web/Features/Admin/Wirecast/Form.razor.cs
@@ -1,21 +1,20 @@
 ﻿using Blazored.Toast.Services;
 using Microsoft.AspNetCore.Components;
 using Microsoft.Extensions.Logging;
-using Microsoft.AspNetCore.Authorization;
 using System.Threading.Tasks;
 using System;
 
 using Page = LivingMessiah.Web.Links.Wirecast;
 
-namespace LivingMessiah.Web.Features.Admin.WirecastFolder;
+namespace LivingMessiah.Web.Features.Admin.Wirecast;
 
-public partial class WirecastEdit
+public partial class Form
 {
-	[Inject] public IShabbatWeekRepository? db { get; set; }
-	[Inject] public ILogger<WirecastEdit>? Logger { get; set; }
+	[Inject] public Data.IRepository? db { get; set; }
+	[Inject] public ILogger<Form>? Logger { get; set; }
 	[Inject] public IToastService? Toast { get; set; }
 
-	public WirecastVM? WirecastVM { get; set; }
+	public WirecastVM? VM { get; set; }
 	public ScratchPad? ScratchPad { get; set; }
 
 	protected int RowCount { get; set; } = 0;
@@ -23,16 +22,25 @@ public partial class WirecastEdit
 	protected override async Task OnInitializedAsync()
 	{
 		Logger!.LogDebug(string.Format("Inside Page: {0}, Class!Method: {1}"
-		, Page.Admin.Index, nameof(WirecastEdit) + "!" + nameof(OnInitializedAsync)));
+		, Page.Admin.Index, nameof(Form) + "!" + nameof(OnInitializedAsync)));
 
 		try
 		{
-			WirecastVM = await db!.GetCurrentWirecast();
-			if (WirecastVM == null)
+		  WirecastQuery wirecastQuery = await db!.GetCurrentWirecast();
+			if (wirecastQuery is null)
 			{
 				string s = $"Wirecast is null after calling {nameof(db.GetCurrentWirecast)}";
 				Logger!.LogWarning(string.Format("...{0}, Sql:{1}", s, db.BaseSqlDump));
 				Toast!.ShowWarning($"...{s}");
+			}
+			else
+			{
+				VM = new WirecastVM()
+				{
+					Id = wirecastQuery!.Id,
+					ShabbatDate = wirecastQuery!.ShabbatDate,
+					WirecastLink = wirecastQuery!.WirecastLink
+				};
 			}
 
 			ScratchPad = await db!.GetScratchPadWireCast();
@@ -46,7 +54,7 @@ public partial class WirecastEdit
 		catch (Exception ex)
 		{
 			Logger!.LogError(ex, string.Format("...Inside catch of {0}"
-				, nameof(WirecastEdit) + "!" + nameof(OnInitializedAsync)));
+				, nameof(Form) + "!" + nameof(OnInitializedAsync)));
 			Toast!.ShowError("An invalid operation occurred reading database, contact your administrator");
 		}
 	}
@@ -56,13 +64,13 @@ public partial class WirecastEdit
 		RowCount = 0;
 		try
 		{
-			RowCount = await db!.UpdateWirecastLink(WirecastVM!.Id, WirecastVM.WirecastLink!);
+			RowCount = await db!.UpdateWirecastLink(VM!.Id, VM.WirecastLink!);
 			Toast!.ShowInfo($"Updated wirecast link, RowCount: {RowCount}");
 		}
 		catch (Exception ex)
 		{
 			Logger!.LogError(ex, string.Format("...Inside catch of {0}"
-				, nameof(WirecastEdit) + "!" + nameof(UpdateWirecastLink)));
+				, nameof(Form) + "!" + nameof(UpdateWirecastLink)));
 			Toast!.ShowError("An invalid operation occurred updating wirecast link, contact your administrator");
 		}
 	}
@@ -78,7 +86,7 @@ public partial class WirecastEdit
 		catch (Exception ex)
 		{
 			Logger!.LogError(ex, string.Format("...Inside catch of {0}"
-			, nameof(WirecastEdit) + "!" + nameof(UpdateScratchPad)));
+			, nameof(Form) + "!" + nameof(UpdateScratchPad)));
 			Toast!.ShowError("An invalid operation occurred updating scratch pad, contact your administrator");
 		}
 	}
