+++ LivingMessiah.Web/Features/Admin/Wirecast/Index.razor
@@ -11,7 +11,7 @@
 <AuthorizeView Roles="@Roles.AdminOrAudiovisual">
 	
 	<Authorized>
-		<WirecastEdit />
+		<Form />
 	</Authorized>
 
 	<NotAuthorized>
