+++ LivingMessiah.Web/Features/Admin/Wirecast/WirecastQuery.cs
@@ -1,0 +1,5 @@
+﻿using System;
+
+namespace LivingMessiah.Web.Features.Admin.Wirecast;
+
+public record WirecastQuery(int Id, DateTime ShabbatDate, string? WirecastLink);
