+++ LivingMessiah.Web/Features/Admin/Wirecast/WirecastVM.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.ComponentModel.DataAnnotations;
 
-namespace LivingMessiah.Web.Features.Admin.WirecastFolder;
+namespace LivingMessiah.Web.Features.Admin.Wirecast;
 
 public class WirecastVM
 {
