--- LivingMessiah.Web/Features/Admin/WirecastFolder/DELETE_RENAME_Wirecast.txt
@@ -1,1 +1,0 @@
-Rename WirecastFolder to just Wirecast
