+++ LivingMessiah.Web/Features/Contacts/Data/ServiceCollectionExtensions.cs
@@ -7,7 +7,7 @@ public static class ServiceCollectionExtensions
 		public static IServiceCollection AddAdminContacts(this IServiceCollection services)
 		{
 				services
-				.AddTransient<IContactRepository, ContactRepository>();
+				.AddTransient<IRepository, Repository>();
 				return services;
 		}
 }
