+++ LivingMessiah.Web/Features/Contacts/Download.cs
@@ -1,7 +1,8 @@
 ﻿using System;
 
-namespace LivingMessiah.Web.Features.Admin.WirecastFolder;
+namespace LivingMessiah.Web.Features.Contacts;
 
+// ToDo: work in progress
 public class Download
 {
 	public bool Selected { get; set; }
@@ -13,6 +14,6 @@ public class Download
 	public string? EMail { get; set; }
 	//public string? Phone { get; set; }
 	public int StatusId { get; set; }
-	public Decimal TotalDonation { get; set; }
-	public Decimal RegistrationFee { get; set; }
+	public decimal TotalDonation { get; set; }
+	public decimal RegistrationFee { get; set; }
 }
