--- LivingMessiah.Web/Features/FeastDayPlanner/NextButton.razor
@@ -1,42 +1,0 @@
-﻿@using FeastDayType = LivingMessiah.Web.Features.Calendar.Enums.FeastDay
-
-@* @if (Next!.IsEndOfEdge) *@
-
-@if (Next is null)
-{
-	<button class="btn btn-light btn-lg disabled">
-		<span class="text-black-50"><i class="fas fa-ban fa-fw fa-2x"></i></span>
-	</button>
-}
-else
-{
-	<button @onclick="@(() => OnClickFilter())"
-					class="btn btn-light btn-lg ">
-		<i class="fas fa-chevron-circle-right fa-fw fa-2x"></i>
-	</button>
-	<p class="mt-1 fw-bold"><small>@Next.PlannerTitle</small></p>
-}
-
-@code {
-	[Parameter, EditorRequired] public required FeastDayType? CurrentFilter { get; set; }
-	[Parameter, EditorRequired] public EventCallback<FeastDayType> OnFilterSelected { get; set; }
-
-	FeastDayType? Next;
-
-	protected override void OnParametersSet()
-	{
-		if (!CurrentFilter!.IsEndOfEdge)
-		{
-			Next = FeastDayType.FromValue(CurrentFilter!.Value + 1);
-		}
-		else
-		{
-			Next = null;
-		}
-	}
-
-	protected void OnClickFilter()
-	{
-		OnFilterSelected.InvokeAsync(Next);
-	}
-}
