+++ LivingMessiah.Web/Features/FeastDayPlanner/OnEdgeButton.razor
@@ -1,0 +1,3 @@
+﻿<button class="btn btn-light btn-lg disabled">
+	<span class="text-black-50"><i class="fas fa-ban fa-fw fa-2x"></i></span>
+</button>
